requirements.txt
swibots
python-decouple
requests